Favor time.Duration over float64 seconds.
This change breaks API compatibility. This is a brand new library and
I am still making decisions.

I have changed the Prepend and Splay methods to take time.Duration
arguments instead of float64 seconds. I felt that choosing seconds as
the base unit was too arbitrary. While it is now a bit more typing, code
becomes more self-documenting. Compare:

	// before
	backoff := retry.Exponential(time.Hour).Splay(300)
	backoff := retry.Exponential(time.Hour).Prepend(10)

	// after
	backoff := retry.Exponential(time.Hour).Splay(time.Minute*5)
	backoff := retry.Exponential(time.Hour).Prepend(time.Second*10)

In addition, this commit adds an Overwrite method, which is like
Prepend but does not preserve the first n args of the original
Strategy, and a Fixed function, which is like Seconds and Milliseconds
but takes time.Duration arguments.